"Update page and dashboard components with new styling and layout changes"
This commit includes changes to the `page.tsx` file, where a new component `BackgroundBeams` is added to the home page, and the layout is updated to remove the fixed positioning. Additionally, there are changes to the `ComponentOne`, `ComponentTwo`, and `ComponentThree` files in the `DashBoard` directory, where the styling of the cards and layout of the tabs have been updated. The `CardDescription` component has also been commented out in some places.